+++ SaavedraCraft/SaavedraCraft.Model/Constructions/BasicContruction.cs
@@ -7,7 +7,7 @@ namespace SaavedraCraft.Model.Constructions
 {
     public abstract class BasicContruction<T> : BasicObject<T>, IConstruction<T>
     {
-        public BasicContruction(string aName, T aComponent, int newI, int newj) : base(aName, aComponent, newI, newj)
+        public BasicContruction(string aName, T aComponent, float newI, float newj) : base(aName, aComponent, newI, newj)
         {
         }
