+++ SaavedraCraft/SaavedraCraft.Model/Interfaces/IObject.cs
@@ -7,14 +7,14 @@ namespace SaavedraCraft.Model.Interfaces
     public interface IObject<T>
     {
         string GetName();
-        int GetCoordI();
-        int GetCoordJ();
+        float GetCoordI();
+        float GetCoordJ();
         int GetWidh();
         int GetHeigh();
         T GetComponentMolde();
         void SetComponentInstanciaReal(T componentReal);
         IObject<T> CloneMe();
-        IObject<T> SetNewIJ(int v1, int v2);
+        IObject<T> SetNewIJ(float v1, float v2);
         void TimeTick(float timedelta);     
         bool isActive();
         void SetActive(bool newValue);
